requirements.txt
openai==0.28
pytest
python-dotenv
